# backend/docs/swagger/swagger.yaml
basePath: /
definitions:
  github_com_yixian-huang_inkless_backend_internal_model.JSONMap:
    additionalProperties: true
    type: object
  github_com_yixian-huang_inkless_backend_internal_service.TranslationState:
    enum:
    - done
    - missing
    - stale
    type: string
    x-enum-varnames:
    - TranslationStateDone
    - TranslationStateMissing
    - TranslationStateStale
  github_com_yixian-huang_inkless_backend_internal_service.ValidationError:
    properties:
      code:
        type: string
      message:
        type: string
      path:
        type: string
    type: object
...
  internal_handler_content.GetDraftResponse:
    properties:
      config:
        $ref: '#/definitions/github_com_yixian-huang_inkless_backend_internal_model.JSONMap'
      pageKey:
        type: string
      publishedVersion:
        type: integer
      updatedAt:
        type: string
      version:
        type: integer
    type: object
...
  internal_handler_content.PublishRequest:
    properties:
      changeNote:
        type: string
      expectedDraftVersion:
        type: integer
    required:
    - expectedDraftVersion
    type: object
  internal_handler_content.PublishResponse:
    properties:
      pageKey:
        type: string
      publishedAt:
        type: string
      publishedVersion:
        type: integer
    type: object
  internal_handler_content.RollbackRequest:
    properties:
      changeNote:
        type: string
    type: object
  internal_handler_content.RollbackResponse:
    properties:
      pageKey:
        type: string
      publishedAt:
        type: string
      publishedVersion:
        type: integer
      sourceVersion:
        type: integer
    type: object
  internal_handler_content.UpdateDraftRequest:
    properties:
      changeNote:
        type: string
      config:
        $ref: '#/definitions/github_com_yixian-huang_inkless_backend_internal_model.JSONMap'
    required:
    - config
    type: object
  internal_handler_content.UpdateDraftResponse:
    properties:
      pageKey:
        type: string
      updatedAt:
        type: string
      version:
        type: integer
    type: object
  internal_handler_content.ValidateRequest:
    properties:
      config:
        $ref: '#/definitions/github_com_yixian-huang_inkless_backend_internal_model.JSONMap'
    required:
    - config
    type: object
  internal_handler_content.ValidateResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/github_com_yixian-huang_inkless_backend_internal_service.ValidationError'
        type: array
      schemaKind:
        type: string
      translationStatus:
        additionalProperties:
          $ref: '#/definitions/github_com_yixian-huang_inkless_backend_internal_service.TranslationState'
        type: object
      valid:
        type: boolean
    type: object
...
host: localhost:8088
info:
  contact: {}
  description: Extensible Inkless CMS API for content, themes, media, plugins, and
    site management.
  title: Inkless CMS API
  version: "1.0"
paths:
...
  /admin/content/{pageKey}/draft:
    get:
      description: Returns draft config for a content_documents pageKey (theme-bound
        home, etc.)
      parameters:
      - description: Page key (home, contact, …)
        in: path
        name: pageKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_content.GetDraftResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: object
            type: object
      security:
      - BearerAuth: []
      summary: Get theme content draft
      tags:
      - Content (Admin)
    put:
      consumes:
      - application/json
      description: Optimistic lock via If-Match draft version; MediaRef url/alt/caption
        must be strings
      parameters:
      - description: Page key
        in: path
        name: pageKey
        required: true
        type: string
      - description: Expected draft version
        in: header
        name: If-Match
        required: true
        type: string
      - description: Draft body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler_content.UpdateDraftRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_content.UpdateDraftResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: object
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: object
            type: object
      security:
      - BearerAuth: []
      summary: Update theme content draft
      tags:
      - Content (Admin)
  /admin/content/{pageKey}/publish:
    post:
      consumes:
      - application/json
      description: Promotes draft to published with expectedDraftVersion lock; invalidates
        content cache
      parameters:
      - description: Page key
        in: path
        name: pageKey
        required: true
        type: string
      - description: Publish parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler_content.PublishRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_content.PublishResponse'
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: object
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            properties:
              error:
                type: object
            type: object
      security:
      - BearerAuth: []
      summary: Publish theme content
      tags:
      - Content (Admin)
  /admin/content/{pageKey}/rollback/{version}:
    post:
      consumes:
      - application/json
      description: Creates a new published version from a historical snapshot (does
        not mutate old versions)
      parameters:
      - description: Page key
        in: path
        name: pageKey
        required: true
        type: string
      - description: Source published version
        in: path
        name: version
        required: true
        type: integer
      - description: Optional change note
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_handler_content.RollbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_content.RollbackResponse'
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: object
            type: object
      security:
      - BearerAuth: []
      summary: Rollback theme content
      tags:
      - Content (Admin)
  /admin/content/{pageKey}/validate:
    post:
      consumes:
      - application/json
      description: Schema + MediaRef checks; returns schemaKind (product-first|corporate|…)
      parameters:
      - description: Page key
        in: path
        name: pageKey
        required: true
        type: string
      - description: Config to validate
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler_content.ValidateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_content.ValidateResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: object
            type: object
      security:
      - BearerAuth: []
      summary: Validate theme content config
      tags:
      - Content (Admin)
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}" for JWT authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"